Moved OpportunityDetailsHeader.tsx and test.tsx to /components/grantor-opportunities, updated imports
