+++ frontend/src/app/[locale]/(base)/grantor/opportunity/[id]/competition/page.tsx
@@ -1,5 +1,4 @@
 import { CompetitionForm } from "src/app/[locale]/(base)/grantor/opportunity/[id]/competition/_components/CompetitionForm";
-import { OpportunityDetailsHeader } from "src/app/[locale]/(base)/grantor/opportunity/[id]/competition/_components/OpportunityDetailsHeader";
 import { ApiRequestError, parseErrorStatus } from "src/errors";
 import withFeatureFlag from "src/services/featureFlags/withFeatureFlag";
 import { getOpportunityForGrantor } from "src/services/fetch/fetchers/opportunitySummaryGrantorFetcher";
@@ -7,6 +6,7 @@ import { getOpportunityForGrantor } from "src/services/fetch/fetchers/opportunit
 import { notFound, redirect } from "next/navigation";
 
 import { UnauthorizedMessage } from "src/components/core/UnauthorizedMessage";
+import { OpportunityDetailsHeader } from "src/components/grantor-opportunities/OpportunityDetailsHeader";
 
 type PageProps = {
   params: Promise<{ id: string; locale: string }>;
