+++ frontend/src/components/grantor-opportunities/OpportunityDetailsHeader.tsx
